Packer: make VM boot time configurable
The VM boot time defines in this case how much we wait until the ISO
image is booted with the empty HDD and the VM reboots for the first
time. After that packer send a <down> key, expecting to be in the GRUB
boot menu, in order to avoid the VM booting for a second time.